Refactored Dataloading in Classes, Viewpoint but not Instances
